pinentry-box/pyproject.toml
[tool.poetry]
name = "pinentry_box"
version = "0.1.0"
description = "A simple program to mesh pinentry"
authors = ["Rizky Maulana Nugraha <lana.pcfre@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
assuan = "^0.2.1b1"
pydevd-pycharm = ">=231.9225.15,<231.9226.0"
pydantic = "^2.5.2"
pyyaml = "^6.0.1"
click = "^8.1.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
pinentry-box = "pinentry_box.cli:main"
